# backend/src/swagger.json
{
   "openapi": "3.0.0",
   "info": {
      "title": "Space App",
      "description": "Final fullstack project for Okten School",
      "version": "1.0.0"
   },
   "servers": [
      {
         "url": "http://localhost:5400/api",
         "description": "Express"
      }
   ],
   "paths": {
...
      "/user": {
         "get": {
            "summary": "get user info",
            "tags": [
               "User"
            ],
            "operationId": "getUserInfo",
            "security": [
               {
                  "Authorization": []
               }
            ],
            "responses": {
               "200": {
                  "description": "Success",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/User"
                        }
                     }
                  }
               },
               "401": {
                  "description": "Unauthorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/UnauthorizedError"
                        }
                     }
                  }
               },
               "default": {
                  "description": "Unexpected errors",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DefaultError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/user/email_change": {
         "post": {
            "summary": "change email request",
            "tags": [
               "User"
            ],
            "operationId": "changeEmailReq",
            "security": [
               {
                  "Authorization": []
               }
            ],
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "type": "object",
                        "properties": {
                           "email": {
                              "type": "string",
                              "example": "johndoe@gmail.com"
                           }
                        }
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Success",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Success"
                        }
                     }
                  }
               },
               "401": {
                  "description": "Unauthorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/UnauthorizedError"
                        }
                     }
                  }
               },
               "default": {
                  "description": "Unexpected errors",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DefaultError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/user/profile_update": {
         "patch": {
            "summary": "update profile",
            "tags": [
               "User"
            ],
            "operationId": "updateProfile",
            "security": [
               {
                  "Authorization": []
               }
            ],
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "type": "object",
                        "properties": {
                           "name": {
                              "type": "number",
                              "example": "John"
                           },
                           "surname": {
                              "type": "string",
                              "example": "Doe"
                           },
                           "username": {
                              "type": "string",
                              "example": "johnny_mnemonick"
                           }
                        }
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Success",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "name": {
                                 "type": "number",
                                 "example": "John"
                              },
                              "surname": {
                                 "type": "string",
                                 "example": "Doe"
                              },
                              "username": {
                                 "type": "string",
                                 "example": "johnny_mnemonick"
                              }
                           }
                        }
                     }
                  }
               },
               "401": {
                  "description": "Unauthorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/UnauthorizedError"
                        }
                     }
                  }
               },
               "default": {
                  "description": "Unexpected errors",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DefaultError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/user/email_new": {
         "patch": {
            "summary": "change email accept",
            "tags": [
               "User"
            ],
            "operationId": "changeEmailAccept",
            "security": [
               {
                  "Authorization": []
               }
            ],
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "type": "object",
                        "properties": {
                           "confirmationToken": {
                              "type": "string",
                              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2M2RmZTA3Nzc0NjNjYjhlODE1NzUyODciLCJlbWFpbCI6ImVsZGFyaW8uYmF5cmFtb3ZAZ21haWwuY29tIiwiaWF0IjoxNjc1NjE2OTkzLCJleHAiOjE2NzU3MDMzOTN9.Mg7ahfZrBwUmqaZLqE4CWR1ALvPqYHmusf10IvKzzi4"
                           }
                        }
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Success",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Success"
                        }
                     }
                  }
               },
               "401": {
                  "description": "Unauthorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/UnauthorizedError"
                        }
                     }
                  }
               },
               "default": {
                  "description": "Unexpected errors",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DefaultError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/user/password_new": {
         "patch": {
            "summary": "change password",
            "tags": [
               "User"
            ],
            "operationId": "changePassword",
            "security": [
               {
                  "Authorization": []
               }
            ],
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "type": "object",
                        "properties": {
                           "currentPassword": {
                              "type": "string",
                              "example": "currpass1"
                           },
                           "newPassword": {
                              "type": "string",
                              "example": "newpass2"
                           }
                        }
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Success",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Success"
                        }
                     }
                  }
               },
               "401": {
                  "description": "Unauthorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/UnauthorizedError"
                        }
                     }
                  }
               },
               "default": {
                  "description": "Unexpected errors",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DefaultError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/user/avatar_upload": {
         "patch": {
            "summary": "upload avatar",
            "tags": [
               "User"
            ],
            "operationId": "uploadAvatar",
            "security": [
               {
                  "Authorization": []
               }
            ],
            "requestBody": {
               "required": true,
               "content": {
                  "multipart/form-data": {
                     "schema": {
                        "type": "object",
                        "properties": {
                           "avatar": {
                              "type": "string",
                              "format": "base64"
                           }
                        }
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Success",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "image": {
                                 "type": "string",
                                 "example": "1677531282980.jpg"
                              }
                           }
                        }
                     }
                  }
               },
               "401": {
                  "description": "Unauthorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/UnauthorizedError"
                        }
                     }
                  }
               },
               "413": {
                  "description": "File size error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "statusCode": {
                                 "type": "number",
                                 "example": 413
                              },
                              "message": {
                                 "type": "string",
                                 "example": "File size must be less than 3 mb"
                              }
                           }
                        }
                     }
                  }
               },
               "422": {
                  "description": "File type error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "statusCode": {
                                 "type": "number",
                                 "example": 422
                              },
                              "message": {
                                 "type": "string",
                                 "example": "Invalid file type"
                              }
                           }
                        }
                     }
                  }
               },
               "default": {
                  "description": "Unexpected errors",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DefaultError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/user/avatar_delete": {
         "patch": {
            "summary": "send file name and delete avatar",
            "tags": [
               "User"
            ],
            "operationId": "deleteAvatar",
            "security": [
               {
                  "Authorization": []
               }
            ],
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "type": "object",
                        "properties": {
                           "fileName": {
                              "type": "string",
                              "example": "1675615755215.jpg"
                           }
                        }
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Success",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Success"
                        }
                     }
                  }
               },
               "401": {
                  "description": "Unauthorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/UnauthorizedError"
                        }
                     }
                  }
               },
               "default": {
                  "description": "Unexpected errors",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/DefaultError"
                        }
                     }
                  }
               }
            }
         }
      },
...
   },
   "components": {
      "securitySchemes": {
         "Authorization": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
         }
      },
      "schemas": {
         "UnauthorizedError": {
            "type": "object",
            "required": [
               "statusCode",
               "message"
            ],
            "properties": {
               "statusCode": {
                  "type": "number",
                  "example": 401
               },
               "message": {
                  "type": "string",
                  "example": "Invalid or expired token"
               }
            }
         },
...
         "DefaultError": {
            "type": "object",
            "properties": {
               "statusCode": {
                  "type": "number",
                  "example": 500
               },
               "message": {
                  "type": "string",
                  "example": "Error message"
               }
            }
         },
         "Success": {
            "type": "object",
            "properties": {
               "message": {
                  "type": "string",
                  "example": "Success"
               }
            }
         },
...
         "User": {
            "type": "object",
            "properties": {
               "name": {
                  "type": "string",
                  "example": "Tom"
               },
               "surname": {
                  "type": "string",
                  "example": "Peterson"
               },
               "username": {
                  "type": "string",
                  "example": "tommy_gun"
               },
               "avatar": {
                  "type": "string",
                  "example": "1677458538916.jpg"
               },
               "momentsCount": {
                  "type": "number",
                  "example": 6
               },
               "plansCount": {
                  "type": "number",
                  "example": 3
               },
               "notesCount": {
                  "type": "number",
                  "example": 1
               }
            }
         }
      }
   }
}